+++ src/pages/Content/scripts/openDfpConsole.ts
