+++ src/pages/Popup/components/adUnits/AdUnitRowComponent.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { IPrebidAdUnit, IPrebidBidWonEventData, IPrebidAdRenderSucceededEventData, IPrebidDetails } from '../../../../inject/scripts/prebid';
+import { IPrebidAdUnit, IPrebidBidWonEventData, IPrebidAdRenderSucceededEventData, IPrebidDetails } from '../../../Content/scripts/prebid';
 import Stack from '@mui/material/Stack';
 import BidChipComponent from './BidChipComponent';
 import AdUnitChipComponent from './AdUnitChipComponent';
