+++ src/pages/Popup/components/adUnits/BidChipComponent.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IPrebidBidWonEventData, IPrebidBid } from '../../../../inject/scripts/prebid';
+import { IPrebidBidWonEventData, IPrebidBid } from '../../../Content/scripts/prebid';
 import Chip from '@mui/material/Chip';
 import Stack from '@mui/material/Stack';
 import JSONViewerComponent from '../../../Shared/JSONViewerComponent';
