+++ src/pages/Popup/components/config/PrebidConfigComponent.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IPrebidConfig } from '../../../../inject/scripts/prebid';
+import { IPrebidConfig } from '../../../Content/scripts/prebid';
 import Typography from '@mui/material/Typography';
 import Avatar from '@mui/material/Avatar';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
