+++ src/pages/Popup/components/config/Server2ServerComponent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Typography from '@mui/material/Typography';
-import { IPrebidConfigS2SConfig } from '../../../../inject/scripts/prebid';
+import { IPrebidConfigS2SConfig } from '../../../Content/scripts/prebid';
 import Card from '@mui/material/Card';
 import CardHeader from '@mui/material/CardHeader';
 import CardContent from '@mui/material/CardContent';
